install vite and material UI

+++ backend/controllers/product.controller.js
@@ -56,6 +56,12 @@ export const updateAProduct = async (req, res) => {
 export const deleteAProduct = async (req, res) => {
     const { id } = req.params;
     console.log(id);
+
+    //add the logic to check whether the id is valid or not
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Invalid Product Id" });
+    }
+    
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Product deleted" });
